+++ src/kandel/geometricKandel/geometricKandelLib.ts
@@ -1,9 +1,10 @@
 import * as ethers from "ethers";
 import { Bigish, typechain } from "../../types";
 
-import { OfferDistribution } from "../kandelDistribution";
+import KandelDistribution, { OfferDistribution } from "../kandelDistribution";
 import GeometricKandelDistribution from "./geometricKandelDistribution";
 import Market from "../../market";
+import GeometricKandelDistributionHelper from "./geometricKandelDistributionHelper";
 
 /** @title Management of a single Kandel instance. */
 class GeometricKandelLib {
@@ -73,18 +74,11 @@ class GeometricKandelLib {
       params.stepSize,
     );
 
-    return {
-      bids: distribution.bids.map((o) => ({
-        index: o.index.toNumber(),
-        gives: bidsOutbound.fromUnits(o.gives),
-        tick: o.tick.toNumber(),
-      })),
-      asks: distribution.asks.map((o) => ({
-        index: o.index.toNumber(),
-        gives: asksOutbound.fromUnits(o.gives),
-        tick: o.tick.toNumber(),
-      })),
-    };
+    return KandelDistribution.mapOffers(distribution, (o, ba) => ({
+      index: o.index.toNumber(),
+      gives: (ba === "bids" ? bidsOutbound : asksOutbound).fromUnits(o.gives),
+      tick: o.tick.toNumber(),
+    }));
   }
 
   public async createFullGeometricDistribution(params: {
@@ -102,6 +96,7 @@ class GeometricKandelLib {
       to: params.pricePoints,
     });
     return new GeometricKandelDistribution(
+      new GeometricKandelDistributionHelper(this.market),
       params.baseQuoteTickIndex0,
       params.baseQuoteTickOffset,
       params.firstAskIndex,